chore(queries): move to specific file
